Fix popover
