+++ src/Pages/Dashboard/ProcessPage.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { ProcessToolbar } from '../../Components/Layouts'
 import Box from '@mui/material/Box';
 import ProcessTable from '../../Components/UI/ProcessTable';
-import { ListModal, TaskModal, SettingsModal } from '../../Components/UI/Modal';
+import { TaskModal, SettingsModal } from '../../Components/UI/Modal';
 import { useFetch, useToggle } from '../../hooks';
 
 const baseUrl = process.env.REACT_APP_BASE_URL
@@ -11,22 +11,11 @@ const ProcessPage = () => {
   const { data, status, error } = useFetch(baseUrl + '/processes')
   const [open, toggle] = useToggle(false)
   const [openSettings, toggleSettings] = useToggle(false)
-  const [openList, toggleList] = useToggle(false)
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+  // const [openList, toggleList] = useToggle(false)
 
   const handleClick = (event: React.MouseEvent<HTMLElement>) => {
     toggle()
   };
-
-  const handleSettingsClick = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const openPopover = Boolean(anchorEl);
 
   if (status === 'error') {
     return <div>
@@ -39,14 +28,8 @@ const ProcessPage = () => {
     <Box>
       {open && <TaskModal open={open} toggleOpen={toggle} />}
       {openSettings && <SettingsModal open={openSettings} toggleOpen={toggleSettings} />}
-      {openList &&
-        <ListModal
-          open={openPopover}
-          toggle={toggleList}
-          anchorEl={anchorEl}
-          handleClose={handleClose}
-        />}
-      <ProcessToolbar toggleOpen={toggleSettings} toggle={handleSettingsClick} />
+
+      <ProcessToolbar toggleOpen={toggleSettings} />
       {data && data.map((process: any, index: string) => (
         <ProcessTable
           process={process}
